Refactors file structure of List compponent for consistency
